Change allocation strategy
The Flat_Pool was not actually helping us in speed at all. Now it takes
the context allocator and wraps it in a tracking allocator so we can
still set a memory limit.

Downside of the tracking allocator is that it creates fat pointers since
it adds the allocation size in the pointer so we can properly track
frees as well.