Greatly reduce allocations
We are now collapsing multiple backtracking state on the same step.
When creating a new backtrack state it will also now search backwards to
the last state that owned either the capture array or group_stack array
and perform a memcmp on it. If it is the same, we just use that array
instead of allocating a new one.