Improve streaming API
Now accepts a `requested` parameter so you know how much to populate at
least. It returns the amount of bytes written to the buffer, so if you
want, you can write more than necessary.

Will also try to keep the buffer small in case you are searching through
GBs of text, so now it returns ranges (start, count) instead of
substrings.

There are some routines to convert those ranges back to substrings if
you have a string of the full input, and we do this by default in the
version that accepts a string.

Unfortunately, this new approach is around 2ms slower than when we had
only the pointer-based string approach.